requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py38h497a2fe_1001
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.5.30=ha878542_0
catalogue=2.0.4=py38h578d9bd_0
certifi=2021.5.30=py38h578d9bd_0
cffi=1.14.6=py38ha65f79e_0
chardet=4.0.0=py38h578d9bd_1
charset-normalizer=2.0.3=pypi_0
click=7.1.2=pyh9f0ad1d_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=3.4.7=py38ha5dfef3_0
cudatoolkit=10.2.89=hfd86e86_1
cupy=9.2.0=py38h14045e6_0
cymem=2.0.5=py38h709712a_2
cython-blis=0.7.4=py38h5c078b8_0
dataclasses=0.8=pyhc8e2a94_1
datasets=1.10.2=pypi_0
dill=0.3.4=pypi_0
elasticsearch=7.13.4=pypi_0
en-core-web-trf=3.1.0=pypi_0
fastrlock=0.6=py38h709712a_1
ffmpeg=4.3=hf484d3e_0
filelock=3.0.12=pypi_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.7.0=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.0.8=pypi_0
idna=3.2=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
jinja2=3.0.1=pyhd8ed1ab_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h2531618_0
markupsafe=2.0.1=py38h497a2fe_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
multiprocess=0.70.12.2=pypi_0
murmurhash=1.0.5=py38h709712a_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
nltk=3.6.2=pypi_0
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h7f98852_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.1=pypi_0
pathy=0.6.0=pyhd8ed1ab_0
pillow=8.3.1=py38h2c7a002_0
pip=21.1.3=py38h06a4308_0
preshed=3.0.5=py38h709712a_1
protobuf=3.17.3=pypi_0
pyarrow=5.0.0=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pydantic=1.8.2=py38h497a2fe_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py38h578d9bd_3
python=3.8.10=h12debd9_8
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch=1.9.0=py3.8_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pypi_0
pyyaml=5.4.1=pypi_0
readline=8.1=h27cfd23_0
regex=2021.7.6=pypi_0
requests=2.26.0=pyhd8ed1ab_0
rouge-score=0.0.4=pypi_0
sacremoses=0.0.45=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.7.1=pypi_0
sentence-transformers=2.0.0=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=52.0.0=py38h06a4308_0
shellingham=1.4.0=pyh44b312d_0
six=1.16.0=pyhd3eb1b0_0
smart_open=5.1.0=pyhd8ed1ab_1
spacy=3.1.1=py38hfc89cab_0
spacy-alignments=0.8.3=pypi_0
spacy-legacy=3.0.8=pyhd8ed1ab_0
spacy-transformers=1.0.3=pypi_0
sqlite=3.36.0=hc218d9a_0
srsly=2.4.1=py38h709712a_0
thinc=8.0.8=py38hfc89cab_0
threadpoolctl=2.2.0=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.3=pypi_0
torchaudio=0.9.0=py38
torchvision=0.10.0=py38_cu102
tqdm=4.61.2=pypi_0
transformers=4.6.1=pypi_0
typer=0.3.2=pyhd8ed1ab_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd8ed1ab_0
wasabi=0.8.2=pyh44b312d_0
wheel=0.36.2=pyhd3eb1b0_0
whoosh=2.7.4=pypi_0
xxhash=2.0.2=pypi_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
